requirements.txt
fastapi
langchain
langchain-community
pypdf
python-dotenv
langchain-groq
nltk
pinecone
pydantic
python-dotenv
sentence-transformers
pandas
openai
uvicorn==0.30.6